+++ src/config/reader/reader_cli.rs
@@ -1,4 +1,3 @@
-use anyhow::anyhow;
 use tokio::fs::File;
 use tokio::io::AsyncReadExt;
 use url::Url;
@@ -29,28 +28,6 @@ impl ConfigReader {
     let mut buffer = Vec::new();
     f.read_to_end(&mut buffer).await?;
     Ok((String::from_utf8(buffer)?, Source::detect(file_path)?))
-  }
-  async fn read_over_url(url: Url) -> anyhow::Result<(String, Source)> {
-    let path = url.path().to_string();
-    let resp = reqwest::get(url).await?;
-    if !resp.status().is_success() {
-      return Err(anyhow!("Read over URL failed with status code: {}", resp.status()));
-    }
-    let source = if let Some(v) = resp.headers().get("content-type") {
-      if let Ok(s) = Source::detect_content_type(v.to_str()?) {
-        s
-      } else {
-        Source::detect(path.trim_end_matches('/'))?
-      }
-    } else {
-      Source::detect(path.trim_end_matches('/'))?
-    };
-    let txt = resp.text().await?;
-    Ok((txt, source))
-  }
-  async fn from_url(url: Url) -> anyhow::Result<Config> {
-    let (st, source) = Self::read_over_url(url).await?;
-    Config::from_source(source, &st)
   }
 }
 #[cfg(test)]
